+++ zinc-compiler/src/semantic/element/type/function/stdlib/tests.rs
@@ -1537,7 +1537,7 @@ fn main() {
 fn error_assets_token_transfer_argument_count_lesser() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, 500 as u248);
+    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1);
 }
 "#;
 
@@ -1560,7 +1560,7 @@ fn main() {
 fn error_assets_token_transfer_argument_count_greater() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, 500 as u248, 1 as u248, 0x666);
+    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, 500 as u248, 0x666);
 }
 "#;
 
@@ -1583,7 +1583,7 @@ fn main() {
 fn error_assets_token_transfer_argument_1_from_expected_u160() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(false, 0x25 as u160, 1, 500 as u248, 1 as u248);
+    std::assets::Token::transfer(false, 0x25 as u160, 1, 500 as u248);
 }
 "#;
 
@@ -1607,7 +1607,7 @@ fn main() {
 fn error_assets_token_transfer_argument_2_to_expected_u160() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(0x42 as u160, false, 1, 500 as u248, 1 as u248);
+    std::assets::Token::transfer(0x42 as u160, false, 1, 500 as u248);
 }
 "#;
 
@@ -1631,7 +1631,7 @@ fn main() {
 fn error_assets_token_transfer_argument_3_token_id_expected_unsigned_integer() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, false, 500 as u248, 1 as u248);
+    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, false, 500 as u248);
 }
 "#;
 
@@ -1655,7 +1655,7 @@ fn main() {
 fn error_assets_token_transfer_argument_4_amount_expected_u248() {
     let input = r#"
 fn main() {
-    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, false, 1 as u248);
+    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, false);
 }
 "#;
 
@@ -1665,30 +1665,6 @@ fn main() {
             function: "transfer".to_owned(),
             name: "amount".to_owned(),
             position: AssetsTokenTransferFunction::ARGUMENT_INDEX_AMOUNT + 1,
-            expected: Type::integer_unsigned(None, zinc_const::bitlength::INTEGER_MAX).to_string(),
-            found: Type::boolean(None).to_string(),
-        }),
-    ))));
-
-    let result = crate::semantic::tests::compile_entry(input);
-
-    assert_eq!(result, expected);
-}
-
-#[test]
-fn error_assets_token_transfer_argument_5_fee_expected_u248() {
-    let input = r#"
-fn main() {
-    std::assets::Token::transfer(0x42 as u160, 0x25 as u160, 1, 500 as u248, false);
-}
-"#;
-
-    let expected = Err(Error::Semantic(SemanticError::Element(ElementError::Type(
-        TypeError::Function(FunctionError::ArgumentType {
-            location: Location::new(3, 78),
-            function: "transfer".to_owned(),
-            name: "fee".to_owned(),
-            position: AssetsTokenTransferFunction::ARGUMENT_INDEX_FEE + 1,
             expected: Type::integer_unsigned(None, zinc_const::bitlength::INTEGER_MAX).to_string(),
             found: Type::boolean(None).to_string(),
         }),